Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -68,7 +68,7 @@ router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, r
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
     creator: req.userData.userId
-    });
+  });
   post.save().then(createdPost => {
     res.status(201).json({
       message: 'Post added successfully',
@@ -93,20 +93,31 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({
-      message: 'Update succesfully'
-    });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: 'Update succesfully'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized'
+      });
+    }
   });
 }
 );
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({
-      message: 'Posts deleted!'
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({
+        message: 'Deletion successful!'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized'
+      });
+    }
   });
 });
